+++ CHANGELOG.md
@@ -81,6 +81,7 @@ CHANGELOG
 - added `intelmq.bots.collectors.calidog.collector_certstream` for collecting certstream data (#1120).
 - added `intelmq.bots.collectors.shodan.collector_stream` for collecting shodan stream data (#1096).
   - Add proxy support.
+  - Fix handling of parameter `countries`.
 
 #### Parsers
 - `bots.parsers.shadowserver`: